# apps/api/docs/openapi.yaml
components:
    schemas:
...
        handler.trainingScheduleDTO:
            properties:
                community_id:
                    nullable: true
                    type: string
                day_of_week:
                    nullable: true
                    type: integer
                end_time:
                    nullable: true
                    type: string
                id:
                    nullable: true
                    type: string
                is_active:
                    nullable: true
                    type: boolean
                location:
                    nullable: true
                    type: string
                start_time:
                    nullable: true
                    type: string
                title:
                    nullable: true
                    type: string
            type: object
...
        handler.upsertTrainingScheduleRequest:
            properties:
                day_of_week:
                    nullable: true
                    type: integer
                end_time:
                    type: string
                is_active:
                    nullable: true
                    type: boolean
                location:
                    nullable: true
                    type: string
                start_time:
                    type: string
                title:
                    type: string
            required:
                - end_time
                - start_time
                - title
            type: object
        handler.userDTO:
            properties:
                avatar_url:
                    nullable: true
                    type: string
                created_at:
                    nullable: true
                    type: string
                email:
                    nullable: true
                    type: string
                full_name:
                    nullable: true
                    type: string
                id:
                    nullable: true
                    type: string
                phone:
                    nullable: true
                    type: string
                updated_at:
                    nullable: true
                    type: string
            type: object
        response.Envelope:
            properties:
                data:
                    nullable: true
                error:
                    $ref: '#/components/schemas/response.ErrorBody'
                meta:
                    $ref: '#/components/schemas/response.Meta'
                success:
                    nullable: true
                    type: boolean
            type: object
        response.ErrorBody:
            nullable: true
            properties:
                code:
                    $ref: '#/components/schemas/response.ErrorCode'
                fields:
                    additionalProperties:
                        type: string
                    nullable: true
                    type: object
                message:
                    nullable: true
                    type: string
            type: object
        response.ErrorCode:
            enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMITED
                - INTERNAL_ERROR
            nullable: true
            type: string
            x-enum-varnames:
                - ErrValidation
                - ErrUnauthorized
                - ErrForbidden
                - ErrNotFound
                - ErrConflict
                - ErrRateLimited
                - ErrInternal
        response.Meta:
            nullable: true
            properties:
                page:
                    nullable: true
                    type: integer
                per_page:
                    nullable: true
                    type: integer
                total:
                    nullable: true
                    type: integer
            type: object
    securitySchemes:
        BearerAuth:
            description: Clerk session JWT, sent as "Bearer <token>".
            in: header
            name: Authorization
            type: apiKey
info:
    contact: {}
    description: Backend API for the Iridescent community hub (events, memberships, announcements, gallery, achievements).
    title: Iridescent Hub API
    version: "1.0"
openapi: 3.0.3
paths:
...
    /schedules:
        get:
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/response.Envelope'
                                    - properties:
                                        data:
                                            items:
                                                $ref: '#/components/schemas/handler.trainingScheduleDTO'
                                            type: array
                                      type: object
                    description: OK
            summary: List active training schedules
            tags:
                - schedules
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/handler.upsertTrainingScheduleRequest'
                description: Schedule
                required: true
                x-originalParamName: body
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/response.Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/handler.trainingScheduleDTO'
                                      type: object
                    description: Created
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: Unprocessable Entity
            security:
                - BearerAuth: []
            summary: Create a training schedule
            tags:
                - schedules
    /schedules/{id}:
        delete:
            parameters:
                - description: Schedule ID
                  in: path
                  name: id
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: Not Found
            security:
                - BearerAuth: []
            summary: Delete a training schedule
            tags:
                - schedules
        patch:
            parameters:
                - description: Schedule ID
                  in: path
                  name: id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/handler.upsertTrainingScheduleRequest'
                description: Schedule
                required: true
                x-originalParamName: body
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/response.Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/handler.trainingScheduleDTO'
                                      type: object
                    description: OK
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: Unprocessable Entity
            security:
                - BearerAuth: []
            summary: Update a training schedule
            tags:
                - schedules
    /users/me:
        get:
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/response.Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/handler.userDTO'
                                      type: object
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: Unauthorized
            security:
                - BearerAuth: []
            summary: Get the authenticated user's profile
            tags:
                - users
    /webhooks/clerk:
        post:
            description: Public but signature-verified (Svix headers). Upserts the local user row — see docs/Schema.md §2.2.
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/response.Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/handler.userDTO'
                                      type: object
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: Unauthorized
            summary: Clerk user.created/user.updated webhook
            tags:
                - users
servers:
    - url: /api/v1